refactor: move all events into a separate file
