+++ eppo_core/src/bandits/mod.rs
@@ -1,7 +1,5 @@
 mod eval;
-mod event;
 mod models;
 
 pub use eval::{get_bandit_action, BanditResult};
-pub use event::BanditEvent;
 pub use models::*;
