+++ eppo_core/src/ufc/assignment.rs
@@ -1,10 +1,6 @@
-use std::collections::HashMap;
-
 use serde::{Deserialize, Serialize};
 
-use crate::Attributes;
-
-use super::eval_details::EvaluationDetails;
+use crate::events::AssignmentEvent;
 
 /// Result of assignment evaluation.
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -233,35 +229,4 @@ impl AssignmentValue {
             _ => None,
         }
     }
-}
-
-/// Represents an event capturing the assignment of a feature flag to a subject and its logging
-/// details.
-#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct AssignmentEvent {
-    /// The key of the feature flag being assigned.
-    pub feature_flag: String,
-    /// The key of the allocation that the subject was assigned to.
-    pub allocation: String,
-    /// The key of the experiment associated with the assignment.
-    pub experiment: String,
-    /// The specific variation assigned to the subject.
-    pub variation: String,
-    /// The key identifying the subject receiving the assignment.
-    pub subject: String,
-    /// Custom attributes of the subject relevant to the assignment.
-    pub subject_attributes: Attributes,
-    /// The timestamp indicating when the assignment event occurred.
-    pub timestamp: String,
-    /// Additional metadata such as SDK language and version.
-    pub meta_data: HashMap<String, String>,
-    /// Additional user-defined logging fields for capturing extra information related to the
-    /// assignment.
-    #[serde(flatten)]
-    pub extra_logging: HashMap<String, String>,
-    /// Evaluation details that could help with debugging the assigment. Only populated when
-    /// details-version of the `get_assigment` was called.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub evaluation_details: Option<EvaluationDetails>,
 }
