+++ eppo_core/src/ufc/mod.rs
@@ -8,6 +8,6 @@ mod rules;
 
 pub mod eval_details;
 
-pub use assignment::{Assignment, AssignmentEvent, AssignmentValue};
+pub use assignment::{Assignment, AssignmentValue};
 pub use eval::{get_assignment, get_assignment_details};
 pub use models::*;
